Cleanup and fixes allowing for difference in v4/5 behavior

+++ src/vitest/coverage-provider.js
@@ -1,13 +1,12 @@
 import { existsSync, readdirSync } from "node:fs";
 import * as fs from "node:fs/promises";
 import path from "node:path";
-import coverage from "istanbul-lib-coverage";
-import libReport from "istanbul-lib-report";
-import reports from "istanbul-reports";
+import { createCoverageMap } from "@vitest/istanbul-lib-coverage";
+import { createAsync, createContext } from "@vitest/istanbul-lib-report";
 import { resolve } from "pathe";
 import c from "picocolors";
-import pm from "picomatch";
 import { coverageConfigDefaults } from "vitest/config";
+import { BaseCoverageProvider } from "vitest/node";
 import { FileCoverageMapService } from "./file-coverage-map-service.js";
 
 /**
@@ -18,7 +17,7 @@ import { FileCoverageMapService } from "./file-coverage-map-service.js";
  *   ResolvedCoverageOptions,
  *   Vitest
  * } from "vitest/node"
- * @import { CoverageMap, CoverageMapData } from "istanbul-lib-coverage"
+ * @import { CoverageMap, CoverageMapData } from "@vitest/istanbul-lib-coverage"
  */
 
 /** @type CoverageProviderModule */
@@ -40,9 +39,13 @@ class JsonSchemaCoverageProvider {
   coverageFilesDirectory = ".json-schema-coverage";
   coverageService = new FileCoverageMapService(path.join(".json-schema-coverage", "maps"));
 
+  #baseProvider = new BaseCoverageProvider();
+
   /** @type CoverageProvider["initialize"] */
   initialize(ctx) {
     this.ctx = ctx;
+    this.#baseProvider.ctx = ctx;
+    this.#baseProvider.createCoverageMap = () => createCoverageMap();
 
     const config = /** @type ResolvedCoverageOptions */ (ctx.config.coverage);
 
@@ -66,6 +69,7 @@ class JsonSchemaCoverageProvider {
         statements: config.thresholds["100"] ? 100 : config.thresholds.statements
       }
     });
+    this.#baseProvider.options = this.options;
 
     const buildScriptPath = path.resolve(import.meta.dirname, "build-coverage-maps.js");
     /** @type string[] */ (ctx.config.globalSetup).push(buildScriptPath);
@@ -111,13 +115,15 @@ class JsonSchemaCoverageProvider {
   }
 
   async onTestFailure() {
-    await this.coverageService.close();
-    await fs.rm(this.coverageFilesDirectory, { recursive: true });
+    if (!this.options.reportOnFailure) {
+      await this.coverageService.close();
+      await fs.rm(this.coverageFilesDirectory, { recursive: true });
+    }
   }
 
   /** @type CoverageProvider["reportCoverage"] */
-  async reportCoverage(coverageMap) {
-    this.#generateReports(/** @type CoverageMap */ (coverageMap) ?? coverage.createCoverageMap());
+  async reportCoverage(coverageMap, reportContext) {
+    await this.#generateReports(/** @type CoverageMap */ (coverageMap) ?? createCoverageMap(), reportContext?.allTestsRun);
 
     // In watch mode we need to preserve the previous results if cleanOnRerun is disabled
     const keepResults = !this.options.cleanOnRerun && this.ctx.config.watch;
@@ -127,40 +133,28 @@ class JsonSchemaCoverageProvider {
     }
   }
 
-  /** @type (coverageMap: CoverageMap) => void */
-  #generateReports(coverageMap) {
-    const context = libReport.createContext({
+  /** @type (coverageMap: CoverageMap, allTestsRun: boolean | undefined) => Promise<void> */
+  async #generateReports(coverageMap, allTestsRun) {
+    const context = createContext({
       dir: this.options.reportsDirectory,
       coverageMap
     });
 
-    if (this.#hasTerminalReporter(this.options.reporter)) {
+    if (this.#baseProvider.hasTerminalReporter(this.options.reporter)) {
       this.ctx.logger.log(c.blue(" % ") + c.dim("Coverage report from ") + c.yellow(this.name));
     }
 
     for (const reporter of this.options.reporter) {
-      // Type assertion required for custom reporters
-      reports
-        .create(/** @type Parameters<typeof reports.create>[0] */ (reporter[0]), {
-          projectRoot: this.ctx.config.root,
-          ...reporter[1]
-        })
-        .execute(context);
+      const report = await createAsync(reporter[0], {
+        projectRoot: this.ctx.config.root,
+        ...reporter[1]
+      });
+      report.execute(context);
     }
 
     if (this.options.thresholds) {
-      this.reportThresholds(coverageMap);
+      await this.#baseProvider.reportThresholds(coverageMap, allTestsRun);
     }
-  }
-
-  /** @type (reporters: ResolvedCoverageOptions["reporter"])=> boolean */
-  #hasTerminalReporter(reporters) {
-    return reporters.some(([reporter]) => {
-      return reporter === "text"
-        || reporter === "text-summary"
-        || reporter === "text-lcov"
-        || reporter === "teamcity";
-    });
   }
 
   /** @type CoverageProvider["onAfterSuiteRun"] */
@@ -171,7 +165,7 @@ class JsonSchemaCoverageProvider {
 
   /** @type CoverageProvider["generateCoverage"] */
   async generateCoverage() {
-    const coverageMap = coverage.createCoverageMap();
+    const coverageMap = createCoverageMap();
 
     for (const file of await fs.readdir(this.coverageFilesDirectory, { recursive: true, withFileTypes: true })) {
       if (!file.isFile()) {
@@ -186,157 +180,6 @@ class JsonSchemaCoverageProvider {
     }
 
     return coverageMap;
-  }
-
-  /**
-   * @typedef {"lines" | "functions" | "statements" | "branches"} Threshold
-   */
-
-  /**
-   * @typedef {{
-   *   coverageMap: CoverageMap
-   *   name: string
-   *   thresholds: Partial<Record<Threshold, number | undefined>>
-   * }} ResolvedThreshold
-   */
-
-  /** @type Set<Threshold> */
-  #THRESHOLD_KEYS = new Set(["lines", "functions", "statements", "branches"]);
-  #GLOBAL_THRESHOLDS_KEY = "global";
-
-  /** @type (coverageMap: CoverageMap) => void */
-  reportThresholds(coverageMap) {
-    const resolvedThresholds = this.#resolveThresholds(coverageMap);
-    this.#checkThresholds(resolvedThresholds);
-  }
-
-  /** @type (coverageMap: CoverageMap) => ResolvedThreshold[] */
-  #resolveThresholds(coverageMap) {
-    /** @type ResolvedThreshold[] */
-    const resolvedThresholds = [];
-    const files = coverageMap.files();
-    const globalCoverageMap = coverage.createCoverageMap();
-
-    const thresholds = /** @type NonNullable<typeof this.options.thresholds> */ (this.options.thresholds);
-    for (const key of /** @type {`${keyof NonNullable<typeof this.options.thresholds>}`[]} */ (Object.keys(thresholds))) {
-      if (key === "perFile" || key === "autoUpdate" || key === "100" || this.#THRESHOLD_KEYS.has(key)) {
-        continue;
-      }
-
-      const glob = key;
-      const globThresholds = resolveGlobThresholds(thresholds[glob]);
-      const globCoverageMap = coverage.createCoverageMap();
-
-      const matcher = pm(glob);
-      const matchingFiles = files.filter((file) => {
-        return matcher(path.relative(this.ctx.config.root, file));
-      });
-
-      for (const file of matchingFiles) {
-        const fileCoverage = coverageMap.fileCoverageFor(file);
-        globCoverageMap.addFileCoverage(fileCoverage);
-      }
-
-      resolvedThresholds.push({
-        name: glob,
-        coverageMap: globCoverageMap,
-        thresholds: globThresholds
-      });
-    }
-
-    // Global threshold is for all files, even if they are included by glob patterns
-    for (const file of files) {
-      const fileCoverage = coverageMap.fileCoverageFor(file);
-      globalCoverageMap.addFileCoverage(fileCoverage);
-    }
-
-    resolvedThresholds.unshift({
-      name: this.#GLOBAL_THRESHOLDS_KEY,
-      coverageMap: globalCoverageMap,
-      thresholds: {
-        branches: this.options.thresholds?.branches,
-        functions: this.options.thresholds?.functions,
-        lines: this.options.thresholds?.lines,
-        statements: this.options.thresholds?.statements
-      }
-    });
-
-    return resolvedThresholds;
-  }
-
-  /** @type (allThresholds: ResolvedThreshold[]) => void */
-  #checkThresholds(allThresholds) {
-    for (const { coverageMap, thresholds, name } of allThresholds) {
-      if (thresholds.branches === undefined && thresholds.functions === undefined && thresholds.lines === undefined && thresholds.statements === undefined) {
-        continue;
-      }
-
-      // Construct list of coverage summaries where thresholds are compared against
-      const summaries = this.options.thresholds?.perFile
-        ? coverageMap.files().map((file) => {
-            return {
-              file,
-              summary: coverageMap.fileCoverageFor(file).toSummary()
-            };
-          })
-        : [{ file: null, summary: coverageMap.getCoverageSummary() }];
-
-      // Check thresholds of each summary
-      for (const { summary, file } of summaries) {
-        for (const thresholdKey of this.#THRESHOLD_KEYS) {
-          const threshold = thresholds[thresholdKey];
-
-          if (threshold === undefined) {
-            continue;
-          }
-
-          /**
-           * Positive thresholds are treated as minimum coverage percentages (X means: X% of lines must be covered),
-           * while negative thresholds are treated as maximum uncovered counts (-X means: X lines may be uncovered).
-           */
-          if (threshold >= 0) {
-            const coverage = summary.data[thresholdKey].pct;
-
-            if (coverage < threshold) {
-              process.exitCode = 1;
-
-              /**
-               * Generate error message based on perFile flag:
-               * - ERROR: Coverage for statements (33.33%) does not meet threshold (85%) for src/math.ts
-               * - ERROR: Coverage for statements (50%) does not meet global threshold (85%)
-               */
-              let errorMessage = `ERROR: Coverage for ${thresholdKey} (${coverage}%) does not meet ${name === this.#GLOBAL_THRESHOLDS_KEY ? name : `"${name}"`} threshold (${threshold}%)`;
-
-              if (this.options.thresholds?.perFile && file) {
-                errorMessage += ` for ${path.relative("./", file).replace(/\\/g, "/")}`;
-              }
-
-              this.ctx.logger.error(errorMessage);
-            }
-          } else {
-            const uncovered = summary.data[thresholdKey].total - summary.data[thresholdKey].covered;
-            const absoluteThreshold = threshold * -1;
-
-            if (uncovered > absoluteThreshold) {
-              process.exitCode = 1;
-
-              /**
-               * Generate error message based on perFile flag:
-               * - ERROR: Uncovered statements (33) exceed threshold (30) for src/math.ts
-               * - ERROR: Uncovered statements (33) exceed global threshold (30)
-               */
-              let errorMessage = `ERROR: Uncovered ${thresholdKey} (${uncovered}) exceed ${name === this.#GLOBAL_THRESHOLDS_KEY ? name : `"${name}"`} threshold (${absoluteThreshold})`;
-
-              if (this.options.thresholds?.perFile && file) {
-                errorMessage += ` for ${path.relative("./", file).replace(/\\/g, "/")}`;
-              }
-
-              this.ctx.logger.error(errorMessage);
-            }
-          }
-        }
-      }
-    }
   }
 }
 
@@ -361,37 +204,6 @@ const resolveCoverageReporters = (configReporters) => {
   }
 
   return resolvedReporters;
-};
-
-/** @type (thresholds: unknown) => ResolvedThreshold["thresholds"] */
-const resolveGlobThresholds = (thresholds) => {
-  if (!thresholds || typeof thresholds !== "object") {
-    return {};
-  }
-
-  if ("100" in thresholds && thresholds["100"] === true) {
-    return {
-      lines: 100,
-      branches: 100,
-      functions: 100,
-      statements: 100
-    };
-  }
-
-  return {
-    lines: "lines" in thresholds && typeof thresholds.lines === "number"
-      ? thresholds.lines
-      : undefined,
-    branches: "branches" in thresholds && typeof thresholds.branches === "number"
-      ? thresholds.branches
-      : undefined,
-    functions: "functions" in thresholds && typeof thresholds.functions === "number"
-      ? thresholds.functions
-      : undefined,
-    statements: "statements" in thresholds && typeof thresholds.statements === "number"
-      ? thresholds.statements
-      : undefined
-  };
 };
 
 export default JsonSchemaCoverageProviderModule;
